tentative inclure T dans NAAC

+++ src/RecuitSimule.h
@@ -16,7 +16,7 @@ public:
 
     Solution correctSolution(Solution solution, const vector<Parking> &vectParkings, const vector<Operation> &vectOperations);
     double fonctionObjectif(Solution solution, const vector<Parking> &vectParkings, const vector<Operation> &vectOperations);
-    void majT(int &acc);
+    void majT(float &acc);
 
     Solution recuitSimule(const vector<Parking> &vectParkings, const vector<Operation> &vectOperation);
     void heatUp();
